refactor: add useUpdateUser hook
